review(LML#536): cleanup pass — drop _SONG_AS_TRACK_VALIDATE_CONCURRENCY, switch fallback to functools.partial, relocate compilations early-exit test
The semaphore that motivated _SONG_AS_TRACK_VALIDATE_CONCURRENCY is gone — the constant is now just the chunk size, hand-coded equal to MAX_SEARCH_RESULTS with the "first chunk satisfies the cap" invariant documented only in prose. Inline MAX_SEARCH_RESULTS at the three _chunked_gather call sites so the invariant is textual at the call, and fold the remaining rationale into the helper's docstring. The test that imported the constant now imports MAX_SEARCH_RESULTS.

Replace the four-line `_fallback_worker` closure with `functools.partial(process_release, skip_self_named_album=False, skip_artist_match_filter=True)` — partial is already imported and used pervasively in this module; the closure was carrying no extra logic.

`_release_info` was unused at both compilations call sites; switch to bare `_`. Add an `assert chunk_size > 0` in `_chunked_gather` so a future misconfiguration fails loudly instead of silently no-op-ing (range step ≤ 0).

Move TestSearchCompilationsEarlyExit out of test_compilation_wave_merge.py (whose docstring scopes it to LML#527 wave-merge gate behavior) into test_orchestrator_helpers.py alongside the parallel TestSearchSongAsTrack early-exit test. Tests now cluster by function rather than scattering by mechanism.
